Move fixture to conftest for auth directory
