+++ tests/integration/web/auth/sudo_test.py
@@ -42,16 +42,3 @@ def other_account(db):
     account.save()
     yield account
     account.delete()
-
-
-@pytest.fixture()
-def session_request(db):
-    """Request object with a real session"""
-    r = RequestFactory()
-    session_request = r.post('/anyurl')
-
-    # use middleware to make session for session_request
-    middleware = SessionMiddleware()
-    middleware.process_request(session_request)
-    session_request.session.save()
-    return session_request
